improve: use JSON version of the events page
Get the list of events from the JSON version of the events page instead
of the HTML version.  This allows use to get _all_ events instead of
just future events.  The JSON also includes a lot more information that
we previously had to get from each event page.  The only information we
need from each event page now is the comments and attendees.

This simplifies the parsing of the event list since we can now use
serde_json to deserialize it.  This also simplfies the parse of the
event page since we only need to parse comments and attendees now.