chore: remove unused min_date from gcal output
Originally, thought it would be needed to implement the original
algorithm which deleted events then re-added.  The algorithm was
reimplented to be patch-or-insert instead of delete-and-add.  This
allows min_date to be implemented on the input side of the gcal output
instead of having to implement min_date internal to the gcal output.