doc: add privary policy
* in order to comply with the Google API Services User Data Policy
* in order to publish the Google Cloud Platform scma-gcal-sync project
* in order to get oauth refresh tokens that don't expire

When Google Cloud Platform projects are in "testing" status, oauth
refresh tokens are only valid for days.  When they expire, the GitHub
Actions event sync workflow breaks.  To get non-expiring refresh tokens,
the project needs to be published.  To get published, it needs to be
verified.  To get verified, it needs to conform to the User Data Policy.